requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0